Add some dependencies to docker container
With this change `arm-vita-eabi-pkg-config` and `make` can be used without adding a lot to the container in terms of size.

CMake I did not add, because it would add an extra 50 megabytes. Let me know if that's still wanted, I can add it.

vdpm is installed in this container, but does not work. This would require some changes to vdpm and the container to make work, though. The story is probably the same for makepkg, but I didn't test that one.