Use root user inside container
I found out that GitHub actions cannot run when not using the root user, so I had to change the strategy for running vita-makepkg a bit. I added sudo and a profile script, which allows the use of `sudo -u user -i vita-makepkg`. It's not perfect, but it works.

The change to the apk binary is no longer needed with this. As you can install packages by default already.